refactor(main): split code from `Application.ts`
