+++ src/window/index.ts
@@ -1,0 +1,3 @@
+export * from './lyric';
+export * from './settings';
+export * from './overlay';
